Plumb outside settings object for ServiceWorkerRegistration#update()
This is similar to crrev.com/c/1861294 but for update(). Script
fetches for update check happen on the browser and are managed
by ServiceWorkerUpdateChecker so pass some information to
ServiceWorkerUpdateChecker too.

This CL has a behavioral change. The 'Referer' header of the request
triggered by update() in a document/service worker will be changed.
Before this CL it was the url of the top-level script but after this
CL it is document/service worker's url. This behavior is consistent
with the spec.

Bug: 937177
Change-Id: I57e9ce7c91a748ad549a49e4e46c7890dcca2903