Reject empty family name in FontFace load call
When document.fonts.load() is called with font specification such as
'1px ""' (containing two quotes) as the argument, CSS font property
parsing was not rejecting this as an invalid font family specification,
even though it was internally known to be empty, leading to a crash in
font face loading.

Add a promise-based WPT test that checks for rejecting such an invalid
font string.

Bug: 1023206
Change-Id: If46aef0c7fcd2e06bc657480b9189857438d8cc1