Resnap containers on layout changes
Per spec, snap containers now resnap to the previous snap target(s)
on relayouts; if the container was not previously snapped then it snaps
to the nearest snap point (includes on initial layout).

This patch only considers targets set on the main thread when
resnapping, as snap targets set on the impl thread are currently not
synced back to main; this behaviour will be implemented in a follow-up.

Bug:866127
Change-Id: I1442e0f464408fb5d9c49c5748a45aec0ed4d166